MDL-63291 calendar: made group event visibility consistent
Previously if you had a capability to see all groups in a course, you
would also see events of all groups in that course, but only if your
calendar was displaying a single course. This change makes it so you can
also see all group events when you are viewing all your courses in the
calendar.