MDL-63119 behat: implemented setting of form fields with same name
It is now possible to set multiple fields with the same label to
different values. To do this just declare the form fields in your
feature file. The first row with a given field name sets the first field
with that label, the second row with the same field name sets the second
field with that label and the nth row sets the nth field.

Example:
When I set the following fields to these values:
      | Event title | Some event |
      | Description | thorough description |
      | Location | |
      | Duration in minutes | |
      | Duration in minutes | 30 |

Here the first 'Duration in minutes' klicks the radio button with that
label while the second one writes '30' into the text field with the same
label.